Fix rust analyzer errors in p2p
